src/requirements.txt
mysqlclient
django
jinja2
django-reversion